HACK: EMIF: BUG00669: Put EMIF in NO_IDLE when DSS is ON
[Problem]:
SYNC Lost between DSS and GFX when CPU_IDLE is enabled
Due to the known EMIF FAST_LOCK issue on OMAP5430 ES1, it
is recommended to keep EMIF in NO_IDLE when Display is ON.
Currently it has been achieved by enabling DSS->EMIF static
dependency. However even with this SD enabled, it's still possible
that EMIF can idle when DSS asserts standby and we do see
underflows with DSS which will lead to SYNC LOST errors.

[Solution]
Explicitly put EMIF in NO_IDLE when DSS is ON and enable
HW_AUTO only when DSS is turned OFF. With this change, HWSUP
flag for EMIF Clock domain can be enabled.

[Test]
Tested Display ON and OFF usecases with CPU_IDLE enabled and
no SYNC Loss messages are seen.

Change-Id: Idd5cd2dac1584399a03081eec5288bc8d7b87f2b
Signed-off-by: Vishwanath Sripathy <vishwanath.bs@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>
Also enabled for OMAP44xx
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>