Regulator: twl6030: Fix twl6030reg_disable func in twl-regulator driver
Currently by using twl6030reg_disable func can not disable any
twl6030 pmic regulators, which is not attached to any group.

This patch fixed a twl6030reg_disable func implementation similarly
to twl6030reg_enable func. That is, the "state_off" will be written
to "vreg_state" register only for those groups to which the regulator
attached now, not for all groups.

Change-Id: I39bc9470d10e327180ddb7e0a0d8df84448ed855
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>